+++ test/clj/rems/api/test_over_http.clj
@@ -15,9 +15,9 @@
 
 (defn- create-test-data []
   (api-key/add-api-key! 42 {:comment "test data"})
-  (test-helpers/create-user! {:eppn "handler"})
-  (test-helpers/create-user! {:eppn "applicant"})
-  (test-helpers/create-user! {:eppn "developer"})
+  (test-helpers/create-user! {:userid "handler"})
+  (test-helpers/create-user! {:userid "applicant"})
+  (test-helpers/create-user! {:userid "developer"})
   (let [wfid (test-helpers/create-workflow! {:handlers ["handler"]})
         form (test-helpers/create-form! nil)
         res-id1 (test-helpers/create-resource! nil)